Create functions for session check and getting user ID
