+++ src/actions/auth/index.ts
@@ -1,3 +1,4 @@
 export * from "./signup";
 export * from "./login";
 export * from "./ressetPassWord";
+export * from "./getUserWithClaims";
