+++ src/app/setup/step-1-user-profile/page.tsx
@@ -1,32 +1,6 @@
+import { authenticated, getUserId } from "@/actions/auth";
 import { UserNameForm } from "@/app/setup/_components";
-import { createClient } from "@/utils/supabase/server";
-import { redirect } from "next/navigation";
 import React from "react";
-
-//Check user session
-const authenticated = async () => {
-  const supabase = await createClient();
-  const { data, error } = await supabase.auth.getClaims();
-
-  if (error || !data) {
-    console.error("Error fetching Claims:", error || "no data");
-
-    redirect("/auth/login");
-  }
-};
-
-//Get user ID
-const getUserId = async () => {
-  const supabase = await createClient();
-  const { data, error } = await supabase.auth.getUser();
-
-  if (error || !data) {
-    console.error("Error fetching Claims:", error || "no data");
-    redirect("/auth/login");
-  }
-
-  return data.user.id;
-};
 
 export default async function SetUserProfile() {
   await authenticated();
